+++ docs/src/components/Sidebar/Sidebar.tsx
@@ -212,8 +212,16 @@ class Sidebar extends React.Component<any, any> {
                 </select>
               </Menu.Item>
             )}
-            <Menu.Item as={NavLink} exact to="/" activeClassName="active">
-              Introduction
+            <Menu.Item>
+              Concepts
+              <Menu.Menu>
+                <Menu.Item as={NavLink} exact to="/" activeClassName="active">
+                  Introduction
+                </Menu.Item>
+                <Menu.Item as={NavLink} exact to="/shorthand-props" activeClassName="active">
+                  Shorthand Props
+                </Menu.Item>
+              </Menu.Menu>
             </Menu.Item>
             <Menu.Item>
               Guides
@@ -238,6 +246,40 @@ class Sidebar extends React.Component<any, any> {
                 </Menu.Item>
               </Menu.Menu>
             </Menu.Item>
+            {process.env.NODE_ENV !== 'production' && (
+              <Menu.Item>
+                Prototypes
+                <Menu.Menu>
+                  <Menu.Item as={NavLink} exact to="/prototype-chat-pane" activeClassName="active">
+                    Chat Pane
+                  </Menu.Item>
+                  <Menu.Item
+                    as={NavLink}
+                    exact
+                    to="/prototype-async-shorthand"
+                    activeClassName="active"
+                  >
+                    Async Shorthand
+                  </Menu.Item>
+                  <Menu.Item
+                    as={NavLink}
+                    exact
+                    to="/prototype-employee-card"
+                    activeClassName="active"
+                  >
+                    Employee Card
+                  </Menu.Item>
+                  <Menu.Item
+                    as={NavLink}
+                    exact
+                    to="/prototype-meeting-options"
+                    activeClassName="active"
+                  >
+                    Meeting Options
+                  </Menu.Item>
+                </Menu.Menu>
+              </Menu.Item>
+            )}
             <Menu.Item active>
               <SemanticUIInput
                 className="transparent inverted icon"