+++ src/components/Chat/index.ts
@@ -1,3 +1,3 @@
 export { default } from './Chat'
 export { default as ChatItem } from './ChatItem'
-export { default as ChatMessage } from './ChatMessage'
+export { default as ChatMessage, IChatMessageProps } from './ChatMessage'