+++ src/components/List/ListItem.tsx
@@ -1,9 +1,15 @@
 import * as React from 'react'
+import * as ReactDOM from 'react-dom'
+
 import * as PropTypes from 'prop-types'
 import { createShorthandFactory, customPropTypes, UIComponent } from '../../lib'
 import ItemLayout from '../ItemLayout'
-import { ListItemBehavior } from '../../lib/accessibility'
+import { listItemBehavior } from '../../lib/accessibility'
 import { Accessibility } from '../../lib/accessibility/interfaces'
+import {
+  FocusableItem,
+  IFocusableItemProps,
+} from '../../lib/accessibility/FocusHandling/FocusableItem'
 import { ComponentVariablesInput, ComponentPartStyle } from '../../../types/theme'
 import { Extendable } from '../../../types/utils'
 
@@ -14,6 +20,7 @@ export interface IListItemProps {
   contentMedia?: any
   content?: any
   debug?: boolean
+  focusableItemProps?: IFocusableItemProps
   header?: any
   endMedia?: any
   headerMedia?: any
@@ -26,7 +33,11 @@ export interface IListItemProps {
   variables?: ComponentVariablesInput
 }
 
-class ListItem extends UIComponent<Extendable<IListItemProps>, any> {
+export interface IListItemState {
+  isHovering: boolean
+}
+
+class ListItem extends UIComponent<Extendable<IListItemProps>, IListItemState> {
   static create: Function
 
   static displayName = 'ListItem'
@@ -61,7 +72,8 @@ class ListItem extends UIComponent<Extendable<IListItemProps>, any> {
     truncateHeader: PropTypes.bool,
 
     /** Accessibility behavior if overridden by the user. */
-    accessibility: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
+    accessibility: PropTypes.func,
+    focusableItemProps: PropTypes.object,
 
     /** Additional CSS styles to apply to the component instance.  */
     styles: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
@@ -72,10 +84,20 @@ class ListItem extends UIComponent<Extendable<IListItemProps>, any> {
 
   static defaultProps = {
     as: 'li',
-    accessibility: ListItemBehavior as Accessibility,
+    accessibility: listItemBehavior as Accessibility,
   }
 
-  state: any = {}
+  constructor(props: IListItemProps) {
+    super(props, null)
+
+    this.state = {
+      isHovering: false,
+    }
+  }
+
+  private itemRef = React.createRef<HTMLElement>()
+
+  private focusableItem = FocusableItem.create(this)
 
   handleMouseEnter = () => {
     this.setState({ isHovering: true })
@@ -83,6 +105,10 @@ class ListItem extends UIComponent<Extendable<IListItemProps>, any> {
 
   handleMouseLeave = () => {
     this.setState({ isHovering: false })
+  }
+
+  componentDidUpdate() {
+    this.focusableItem.tryFocus(ReactDOM.findDOMNode(this.itemRef.current) as HTMLElement)
   }
 
   renderComponent({ ElementType, classes, accessibility, rest, styles }) {
@@ -139,6 +165,7 @@ class ListItem extends UIComponent<Extendable<IListItemProps>, any> {
         headerCSS={headerCSS}
         headerMediaCSS={headerMediaCSS}
         contentCSS={contentCSS}
+        ref={this.itemRef}
         {...accessibility.attributes.root}
         {...rest}
       />