+++ src/themes/teams/components/Icon/iconStyles.ts
@@ -20,10 +20,7 @@ const sizes = new Map([
 ])
 
 const getDefaultFontIcon = (iconName: string) => {
-  const fontFamily = iconName && iconName.includes('outline') ? 'outline-icons' : 'Icons'
-  const content = (iconName && `'\\${fontAwesomeIcons(iconName)}'`) || '?'
-
-  return { content, fontFamily }
+  return callable(fontAwesomeIcons(iconName).icon)()
 }
 
 const getSize = size => `${sizes.get(size)}em`
@@ -126,6 +123,8 @@ const iconStyles: IComponentPartStylesInput<IIconProps, any> = {
   g: getSvgStyle('g'),
 
   /* additional SVG styles for different paths could be added/used in the same way */
+  path: getSvgStyle('path'),
+  secondaryPath: getSvgStyle('secondaryPath'),
 }
 
 export default iconStyles